// server/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  adminFlags: typeof adminFlags;
  files: typeof files;
  http: typeof http;
  map: typeof map;
  reports: typeof reports;
  seedAdmin: typeof seedAdmin;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;